Add support for product variants
To watch only a single variant, specify foo@123 in the TOML array. The number can be found in the URL after selecting the desired variant on the product page.

This also renames the TOML key from 'items' to 'products'.

Additionally, if you do not specify a product variant then you will not be notified of any variant's availability.